Refactor rbac package
